requirements.txt
ffmpeg
ipympl
ipykernel
librosa
matplotlib
mpl_pan_zoom
mpl_point_clicker
numpy
pandas
PeakUtils
# plotly
#scikit_learn
scikit_maad
scipy
sympy